delete review done

+++ server/resolvers/deleteReview.js
@@ -1,24 +1,30 @@
 const { ApolloError, AuthenticationError } = require("apollo-server");
+const { differenceInCalendarQuarters } = require("date-fns");
 
 const { Review, Event, User } = require("../models");
 
 const deleteReview = async (_, { reviewId }, { user }) => {
   try {
-    console.log(user);
     if (user) {
-      const deleteReviewFromDatabase = await Review.findByIdAndDelete(reviewId);
-      const { postedBy } = deleteReviewFromDatabase;
-      const { eventId } = deleteReviewFromDatabase;
-      await User.findOneAndUpdate(
-        postedBy,
-        { $pull: { reviews: reviewId } },
-        { new: true }
-      );
-      await Event.findOneAndUpdate(
-        eventId,
-        { $pull: { reviews: { reviewId } } },
-        { new: true }
-      );
+      const { _id: userId } = user;
+
+      const { eventId, postedBy } = Review.findById(reviewId);
+      if (userId === postedBy) {
+        await Review.findByIdAndDelete(reviewId);
+
+        await User.findOneAndUpdate(
+          userId,
+          { $pull: { reviews: reviewId } },
+          { new: true }
+        );
+        await Event.findOneAndUpdate(
+          eventId,
+          { $pull: { reviews: { reviewId } } },
+          { new: true }
+        );
+      } else {
+        console.log("Not ur review");
+      }
     } else {
       throw new AuthenticationError(
         "You must be logged in to delete a review."
